Create mod_offline_combat_simulator_csnosorog.ltx
Sets the squads for Warfare. Anatoly's, ISKRA and Hatchet's should be fairly tanky in offline combat. Anatoly with the rookies should be a hard nut to crack and force the player to participate to defeat them.
The fishing hamlet and Agroprom factory have factions assigned, and all custom squads are stationary (not controllable by player).